# workflow/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Finternet UNITS APIs Documentation",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
...
        "/v1/actor/create": {
            "post": {
                "description": "Creates a new user actor, generates and stores the reproducible DID from the master public key, and links them.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Register a new actor",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.ApiRequest_RegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_RegistrationSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request, such as a missing required field",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    },
                    "409": {
                        "description": "Conflict, an identifier is already in use",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    }
                }
            }
        },
        "/v1/actor/forgotPassword": {
            "post": {
                "description": "Initiates the password reset flow for a given email address. A reset link will be sent to the user's email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Forgot password",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.ApiRequest_ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Success"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    }
                }
            }
        },
        "/v1/actor/getProfile": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the full actor profile for the authenticated actor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Get current actor's profile",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.ApiRequest_Empty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_ActorProfile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    }
                }
            }
        },
        "/v1/actor/login": {
            "post": {
                "description": "Authenticates an actor and returns JWT tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Login actor",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.ApiRequest_LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    }
                }
            }
        },
        "/v1/actor/resolve": {
            "post": {
                "description": "Publicly resolves a universal identifier to its master public key and DID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Resolve a universal identifier",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.ApiRequest_ResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_ResolveResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    }
                }
            }
        },
        "/v1/actor/signout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Logs out the user from auth provider and revokes all their active sessions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Actor signout",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.ApiRequest_Empty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Success"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    }
                }
            }
        },
        "/v1/actor/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an authenticated user to update their profile information, such as name and phone number. The actor to update is determined by the session token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Update actor profile",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validation.ApiRequest_UpdateActorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_ActorProfile"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    },
                    "409": {
                        "description": "Conflict, the new phone number is already in use",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse_Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.ApiResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/response.ApiResponseParams"
                },
                "response": {},
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "response.ApiResponseParams": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "errmsg": {
                    "type": "string"
                },
                "msgid": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "status": {
                    "type": "string",
                    "example": "successful"
                }
            }
        },
        "response.ApiResponse_ActorProfile": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/response.ApiResponseParams"
                },
                "response": {
                    "$ref": "#/definitions/response.ActorProfile"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "response.ApiResponse_Error": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/response.ApiResponseParams"
                },
                "response": {},
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "response.ApiResponse_RegistrationSuccess": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/response.ApiResponseParams"
                },
                "response": {
                    "$ref": "#/definitions/response.RegistrationSuccessResponse"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "response.ApiResponse_ResolveResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/response.ApiResponseParams"
                },
                "response": {
                    "$ref": "#/definitions/response.ResolveResponse"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
        "response.ApiResponse_Success": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/response.ApiResponseParams"
                },
                "response": {
                    "$ref": "#/definitions/response.SuccessResponse"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "v1"
                }
            }
        },
...
        "response.RegistrationSuccessResponse": {
            "type": "object",
            "properties": {
                "did": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "message": {
                    "type": "string",
                    "example": "Actor registered successfully"
                }
            }
        },
...
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "validation.ApiRequest_Empty": {
            "type": "object",
            "required": [
                "id",
                "ts",
                "ver"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/validation.RequestParams"
                },
                "request": {
                    "type": "object"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "5.0.0"
                }
            }
        },
        "validation.ApiRequest_ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "id",
                "ts",
                "ver"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/validation.RequestParams"
                },
                "request": {
                    "$ref": "#/definitions/validation.ForgotPasswordRequest"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "5.0.0"
                }
            }
        },
        "validation.ApiRequest_LoginRequest": {
            "type": "object",
            "required": [
                "id",
                "ts",
                "ver"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/validation.RequestParams"
                },
                "request": {
                    "$ref": "#/definitions/validation.LoginRequest"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "5.0.0"
                }
            }
        },
        "validation.ApiRequest_RegistrationRequest": {
            "type": "object",
            "required": [
                "id",
                "ts",
                "ver"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/validation.RequestParams"
                },
                "request": {
                    "$ref": "#/definitions/validation.RegistrationRequest"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "5.0.0"
                }
            }
        },
        "validation.ApiRequest_ResolveRequest": {
            "type": "object",
            "required": [
                "id",
                "ts",
                "ver"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/validation.RequestParams"
                },
                "request": {
                    "$ref": "#/definitions/validation.ResolveRequest"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "5.0.0"
                }
            }
        },
        "validation.ApiRequest_UpdateActorRequest": {
            "type": "object",
            "required": [
                "id",
                "ts",
                "ver"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "api.actor.create"
                },
                "params": {
                    "$ref": "#/definitions/validation.RequestParams"
                },
                "request": {
                    "$ref": "#/definitions/validation.UpdateActorRequest"
                },
                "ts": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "ver": {
                    "type": "string",
                    "example": "5.0.0"
                }
            }
        },
...
        "validation.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "actor@example.com"
                }
            }
        },
...
        "validation.RegistrationRequest": {
            "type": "object",
            "required": [
                "email",
                "entityType",
                "firstName",
                "lastName",
                "masterPublicKey",
                "password",
                "universalIdentifier"
            ],
            "properties": {
                "countryOfIncorporation": {
                    "description": "Business fields",
                    "type": "string",
                    "example": "US"
                },
                "countryOfResidence": {
                    "type": "string",
                    "example": "US"
                },
                "email": {
                    "type": "string",
                    "example": "actor@example.com"
                },
                "entityType": {
                    "type": "string",
                    "enum": [
                        "Individual",
                        "Business"
                    ],
                    "example": "Individual"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "masterPublicKey": {
                    "type": "string",
                    "example": "-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA...\n-----END PUBLIC KEY-----"
                },
                "nationality": {
                    "description": "Individual fields",
                    "type": "string",
                    "example": "US"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "password123"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "universalIdentifier": {
                    "type": "string",
                    "example": "actor123"
                }
            }
        },
        "validation.RequestParams": {
            "type": "object",
            "properties": {
                "msgid": {
                    "type": "string",
                    "example": "2f7c3a4e-8bcb-4b44-8cf6-84cc8f1d2a87"
                }
            }
        },
...
    }
}